% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchloc.R
\name{searchloc}
\alias{searchloc}
\title{function for search SENAMHI stations information by coordinates.}
\usage{
searchloc(x, stations, dist_max)
}
\arguments{
\item{x}{numeric; numeric vector of coordinates (latitude and longitude)}

\item{stations}{stations information of SENAMHI web scraping}

\item{dist_max}{numeric, value of maximum distance from stations to point (kms)}
}
\description{
function for search SENAMHI stations information by coordinates (lat and lon) inside SENAMHI data web.
}
\examples{

x <- c(-13.18, -74.23)
dist_max <- 50
stations <- stations()

searchloc(x, stations, dist_max)

}
\author{
Geomar Perales Apaico
}
