% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchname.R
\name{searchname}
\alias{searchname}
\title{function for search SENAMHI stations information by station name.}
\usage{
searchname(x, stations)
}
\arguments{
\item{x}{character; optional character vector to filter results by station name.}

\item{stations}{stations information of SENAMHI web scraping}
}
\description{
function for search SENAMHI stations information by station name inside SENAMHI data web.
}
\examples{

x <- "CORD"
stations <- stations()

searchname(x, stations)

}
\author{
Geomar Perales Apaico
}
