% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhiwsdir.R
\name{senamhiwsdir}
\alias{senamhiwsdir}
\title{function for get records of SENAMHI stations  in a specific folder.}
\usage{
senamhiwsdir(x, stations, dir = "senamhi")
}
\arguments{
\item{x}{a vector from SENAMHI stations code}

\item{stations}{stations information of SENAMHI web scraping}

\item{dir}{folder to save SENAMHI records}
}
\description{
function for get record of SENAMHI stations from 01-01-2016 to 31-12-2022 and
save data files and save dataset files in a specific folder.
}
\examples{

x <- c("114123", "114030")
stations <- stations()
senamhiwsdir(x, stations, dir = "senamhi")

}
\author{
Geomar Perales Apaico
}
