% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{function for get}
\usage{
stations()
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
function for processing of pisco daily values to a monthly average, apply to precipitation,
temperature and evapotranspiration.
}
\author{
Geomar Perales Apaico
}
