% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhinfo.R
\name{senamhinfo}
\alias{senamhinfo}
\title{function for get information about stations.}
\usage{
senamhinfo()
}
\description{
function for get information (coordinates, station code and height) about stations.
}
\author{
Geomar Perales Apaico
}
