% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{function for get all stations information of SENAMHI web scraping}
\usage{
stations()
}
\description{
function for get all stations information (coordinates and altitude) of SENAMHI web scraping.
}
\examples{

stations <- stations()
}
\author{
Geomar Perales Apaico
}
